Move `SwiftDataExerciseStore` to production
