+++ WorkoutTrack/Tests/WorkoutTrackTests/Exercises Cache/SwiftDataExerciseStoreTests.swift
@@ -9,31 +9,6 @@ import XCTest
 @testable import WorkoutTrack
 import SwiftData
 
-@ModelActor
-final actor SwiftDataExerciseStore {
-    
-    func insert(_ exercise: CustomExercise) {
-        let entity = ExerciseEntity(id: exercise.id, name: exercise.name, category: exercise.category)
-        modelContext.insert(entity)
-    }
-    
-    func retrieve(by query: ExerciseQuery) throws -> [CustomExercise] {
-        var descriptor = FetchDescriptor<ExerciseEntity>()
-        if let predicate = query.predicate {
-            descriptor.predicate = predicate
-        }
-        
-        if let sort = query.sortDescriptor {
-            descriptor.sortBy = [sort]
-        } else {
-            descriptor.sortBy = [SortDescriptor(\.name, order: .forward)]
-        }
-        
-        return try modelContext.fetch(descriptor).toModels()
-    }
-}
-
-
 final class SwiftDataExerciseStoreTests: XCTestCase {
     
     func test_retrieve_all_deliversEmptyOnEmptyDatabase() async throws {
